3D_X_ray_angiography/requirements.txt
numpy~=1.19.0
pypng
pydicom~=2.0.0
matplotlib~=3.2.2
geneticalgorithm~=1.0.1
pypcd~=0.1.3
scipy~=1.5.1
opencv-python~=4.3.0.36
open3d~=0.10.0.1
geomdl~=5.2.10
torch~=1.2.0
numba~=0.50.1